% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issm.R, R/methods.R
\name{add_arma.tsissm.component}
\alias{add_arma.tsissm.component}
\alias{add_arma}
\title{ARMA Component}
\usage{
\method{add_arma}{tsissm.component}(x, order = c(0, 0), ar = 0, ma = 0, mu = 0, ...)

add_arma(x, ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{order}{the ar and ma orders.}

\item{ar}{a vector of ar coefficients.}

\item{ma}{a vector of ma coefficients.}

\item{mu}{the mean parameter (defaults to zero) of the ARMA process.}

\item{...}{additional parameters.}
}
\value{
An object of class tsissm.component updated with the ARMA
component.
}
\description{
ARMA Component
}
