% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issm.R, R/methods.R
\name{add_regressor.tsissm.component}
\alias{add_regressor.tsissm.component}
\alias{add_regressor}
\title{Regressor Component}
\usage{
\method{add_regressor}{tsissm.component}(x, xreg = NULL, pars = NULL, ...)

add_regressor(x, ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{xreg}{a matrix of regressors.}

\item{pars}{regressors coefficients.}

\item{...}{additional parameters.}
}
\value{
An object of class tsissm.component updated with the regressor
components.
}
\description{
Regressor Component
}
