% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issm.R, R/methods.R
\name{add_transform.tsissm.component}
\alias{add_transform.tsissm.component}
\alias{add_transform}
\title{Transform}
\usage{
\method{add_transform}{tsissm.component}(x, method = "box-cox", lambda = 1, lower = 0, upper = 1, ...)

add_transform(x, ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{method}{a valid transform.}

\item{lambda}{the Box-Cox parameter.}

\item{lower}{the lower bound for the logit transform.}

\item{upper}{the upper bound for the logit transform.}

\item{...}{additional parameters.}
}
\value{
An object of class tsissm.component updated with the transformation.
}
\description{
Transform
}
\details{
The inverse transform is applied to the simulated series based on
the \code{\link{tstransform}} method.
}
