% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssimulator-package.R
\docType{package}
\name{tssimulator-package}
\alias{tssimulator}
\alias{tssimulator-package}
\title{tssimulator: Time Series Simulation}
\description{
Allows the easy generation of random time series under a number of data generating processes and mixtures.
}
\examples{
library(magrittr)
set.seed(105)
sim <- initialize_simulator(rnorm(12*20, 0, 1), sampling = "months")
sim <- sim \%>\% add_polynomial(order = 2, alpha = 0.1, beta = 0.02, l0 = 100, b0 = 0.25) \%>\%
add_seasonal(frequency = 12, gamma = 0.01, init_harmonics = 4, init_scale = 3)
# plot(sim, lwd = 2, ylab = "")
# plot(sim, type = "components")
}
\keyword{internal}
