% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssimulator-package.R
\docType{package}
\name{tssimulator-package}
\alias{tssimulator}
\alias{tssimulator-package}
\title{tssimulator: Time Series Simulation}
\description{
Allows the easy generation of random time series under a number of data generating processes and mixtures.
}
\keyword{internal}
